refactor: fix incorrect internal imports
Importing from src/blah/blah causes problems for api-extractor
