+++ src/components/chip-set/chip.types.ts
@@ -1,4 +1,4 @@
-import { Icon } from '../../interface';
+import { Icon } from '../../global/shared-types/icon.types';
 
 /**
  * @public
