+++ src/components/file-viewer/file-viewer.tsx
@@ -14,7 +14,7 @@ import translate from '../../global/translations';
 import { detectExtension } from './extension-mapping';
 import { Fullscreen } from './fullscreen';
 import { FileType, OfficeViewer } from './file-viewer.types';
-import { LimelMenuCustomEvent } from 'src/components';
+import { LimelMenuCustomEvent } from '../../components';
 
 /**
  * This is a smart component that automatically detects
